fix(compiler): centralize compilation context creation
