+++ packages/compiler/src/utils/testUtils.ts
@@ -1,78 +1,26 @@
 import { BlockType } from '@8f4e/compiler-spec';
 
-import type { BlockStack, CompilationContext } from '@8f4e/compiler-spec';
+import { createCompilationContext } from '../semantic/createCompilationContext';
 
-function getBlockContextFlags(blockStack: BlockStack) {
-	return {
-		insideModuleBlock: blockStack.some(block => block.blockType === BlockType.MODULE),
-		insideFunctionBlock: blockStack.some(block => block.blockType === BlockType.FUNCTION),
-		insideGenericBlock: blockStack.some(block => block.blockType === BlockType.BLOCK),
-		insideLoopBlock: blockStack.some(block => block.blockType === BlockType.LOOP),
-		insideConditionBlock: blockStack.some(block => block.blockType === BlockType.CONDITION),
-		insideConstantsBlock: blockStack.some(block => block.blockType === BlockType.CONSTANTS),
-		insideMapBlock: blockStack.some(block => block.blockType === BlockType.MAP),
-	};
-}
+import type { CompilationContext } from '@8f4e/compiler-spec';
 
 export default function createInstructionCompilerTestContext(
 	overrides: Partial<CompilationContext> = {}
 ): CompilationContext {
-	const base: CompilationContext = {
+	return createCompilationContext({
+		...overrides,
 		namespace: {
-			namespaces: {},
-			memory: {},
-			consts: {},
 			moduleName: 'test',
+			...overrides.namespace,
 		},
-		locals: {},
-		internalResources: {},
-		internalAllocator: {
-			nextByteAddress: 0,
-		},
-		stack: [],
-		blockStack: [
+		blockStack: overrides.blockStack ?? [
 			{
 				blockType: BlockType.MODULE,
 				expectedResultIsInteger: false,
 				hasExpectedResult: false,
 			},
 		],
-		insideModuleBlock: true,
-		insideFunctionBlock: false,
-		insideGenericBlock: false,
-		insideLoopBlock: false,
-		insideConditionBlock: false,
-		insideConstantsBlock: false,
-		insideMapBlock: false,
-		startingByteAddress: 0,
-		currentModuleNextWordOffset: 0,
-		currentModuleWordAlignedSize: 0,
-		currentMemoryIndex: 0,
-		memoryRegions: [],
-		byteCode: [],
-		codeBlockId: 'test',
-		codeBlockType: 'module',
-	};
-
-	const context = {
-		...base,
-		...overrides,
-		namespace: {
-			...base.namespace,
-			...overrides.namespace,
-		},
-	};
-
-	const blockContextFlags = getBlockContextFlags(context.blockStack);
-
-	return {
-		...context,
-		insideModuleBlock: overrides.insideModuleBlock ?? blockContextFlags.insideModuleBlock,
-		insideFunctionBlock: overrides.insideFunctionBlock ?? blockContextFlags.insideFunctionBlock,
-		insideGenericBlock: overrides.insideGenericBlock ?? blockContextFlags.insideGenericBlock,
-		insideLoopBlock: overrides.insideLoopBlock ?? blockContextFlags.insideLoopBlock,
-		insideConditionBlock: overrides.insideConditionBlock ?? blockContextFlags.insideConditionBlock,
-		insideConstantsBlock: overrides.insideConstantsBlock ?? blockContextFlags.insideConstantsBlock,
-		insideMapBlock: overrides.insideMapBlock ?? blockContextFlags.insideMapBlock,
-	};
+		codeBlockId: overrides.codeBlockId ?? 'test',
+		codeBlockType: overrides.codeBlockType ?? 'module',
+	});
 }
